run.py: stop hiding sandbox and recommend --debug on sandbox failure
As discussed in #3948 and #3992, logging commands failing in a sandbox
without ever mentioning that a sandbox was in use is misleading and very
time-consuming; notably because paths look real but they are not.

To remove this confusion and cost, simply add one short line mentioning
that a sandbox was involved when the failure happened + recommend to
re-run with --debug to display bind mounts and other important
information.

Also separate the command and error message on different lines for
better readability and re-usability.

Just for the record: before commit 7e7a3c71696d ("Don't log sandbox for
every command"), `run.py` was logging sandbox commands every time which
was too verbose and unreadable. That commit said "... but still log the
full sandbox if a command fails" and it did that do in `__init__.py` -
but not always in `run.py`. After that commit, the sandbox was logged by
`run.py` only in --debug mode and became completely invisible otherwise.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>